deepseekec638/requirements.txt
streamlit
langchain_core
langchain-community
langchain_ollama
pdfplumber